Reintroduced accidentally deleted files. Modified css slightly.
I added back in files that I had accidentally removed when I was first setting things up. Without these (for example .babelrc) it's impossible to run the application using the instructions provided on electron-react-boilerplate (ERB).

I also began fooling around with removing the sample project that currently comes when you clone ERB. Specifically I changed the dimensions of the main window to 1200px x 600px so that it will match the dimensions of the Pi. This may proved to be unnecessary down the line, but for now I've left it in.