Changed the value of the "kiosk" property to false
Changed the value of the "kiosk" property to false. When it was set to true and I launched the application I found that the only way to see other applications (like my IDE) was to quit Electron. It might be possible for our own dev purposes to put in a key binding which would toggle Kiosk mode during development, but for now it's easiest just to keep it set to false.