Added more features to DescribeInput component
Component now displays actively updated character count
Component cannot effectively gain focus (it is set to blur once it is selected)
Component now triggers display of keyboard whenever selected
Component now receives value of input field as a prop passed down by the Describe component.
The Describe component also now maintains the string value of the description the user has entered as part of the component's state, which more easily allows for it to be passed to other components and to keep track of the length of the string for use in updating the character count.