Initial layout of draw page
This is intended to include a general layout of the different aspects of the drawing page. For the most part functionality has not been added with the intention of getting something up and running first and focusing on making it pretty and completely functional later. This way we can start generating images which can be saved to remote storage and entered into database records so that those parts of the project can be added and tested.