Save UID of Current Player into component state
This way when the DrawPage.js component is loaded, it will retrieve this
value along with the string value for the current description and pass
them as props to the Draw component. When the Draw component is completed
the UID, description and drawing's URL will be pushed into the db for use
in the next round.